language-plugins/python/src/index.py
import ast


class PythonProcessor:
    """A powerful Python code processor for parsing, optimizing, and generating Python code."""

    def parse_code(self, code: str) -> ast.AST:
        """Parse Python code into an Abstract Syntax Tree (AST)."""
        try:
            return ast.parse(code)
        except SyntaxError as e:
            raise ValueError(f"Error parsing code: {e}")

    def optimize_code(self, tree: ast.AST) -> ast.AST:
        """Perform optimizations on the AST."""
        # Currently, this is a placeholder for optimizations.
        return tree

    def generate_code(self, tree: ast.AST) -> str:
        """Generate Python code from an AST."""
        try:
            return ast.unparse(tree).strip() + "\n"
        except Exception as e:
            raise ValueError(f"Error generating code: {e}")
